move command files
